# keydivide/backend/apidoc.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Интернет-магазина эргономичных клавиатур",
      "description": "REST API для управления каталогом клавиатур, свитчами, заказами и пользователями",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:5000",
        "description": "Сервер разработки"
      }
    ],
    "paths": {
...
        "/admin/switches": {
            "post": {
              "summary": "Добавление нового свитча",
              "description": "Создание нового типа свитча для кастомизации клавиатур",
              "tags": ["Admin"],
              "security": [
                {
                  "bearerAuth": []
                }
              ],
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "required": ["name", "type", "price"],
                      "properties": {
                        "name": {
                          "type": "string",
                          "description": "Название свитча",
                          "minLength": 1,
                          "maxLength": 255
                        },
                        "type": {
                          "type": "string",
                          "description": "Тип свитча (линейные, тактильные, линейные тихие)"
                        },
                        "description": {
                          "type": "string",
                          "description": "Описание свитча"
                        },
                        "actuation_force": {
                          "type": "string",
                          "description": "Сила срабатывания (в граммах)"
                        },
                        "travel_distance": {
                          "type": "string",
                          "description": "Дистанция хода (в мм)"
                        },
                        "price": {
                          "type": "number",
                          "minimum": 0,
                          "description": "Цена свитча"
                        },
                        "image_url": {
                          "type": "string",
                          "description": "URL изображения свитча"
                        },
                        "in_stock": {
                          "type": "boolean",
                          "description": "Наличие на складе",
                          "default": true
                        }
                      }
                    }
                  }
                }
              },
              "responses": {
                "201": {
                  "description": "Свитч успешно создан",
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/Switch"
                      }
                    }
                  }
                },
                "400": {
                  "description": "Невалидные данные свитча",
                  "content": {
                    "application/json": {
                      "schema": {
                        "$ref": "#/components/schemas/Error"
                      }
                    }
                  }
                },
                "401": {
                  "description": "Требуется аутентификация администратора"
                }
              }
            }
        },
...
    },
    "components": {
      "schemas": { 
...
        "Switch": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "actuation_force": {
              "type": "string"
            },
            "travel_distance": {
              "type": "string"
            },
            "price": {
              "type": "number",
              "format": "float"
            },
            "image_url": {
              "type": "string"
            },
            "in_stock": {
              "type": "boolean"
            }
          }
        },
...
        "Error": {
          "type": "object",
          "properties": {
            "error": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
...
  }